Create Solution.java
this solution will also work for unicode character as we are using the Hashtable instead of the int array. However, if the test case are guarantee to have ONLY a-z then the int array will definitely be more efficient. In addition, instead of using the toCharArray(), we could instead use the s.charAt(i) to get character out from the string which will use a lot less space as toCharArray makes a copy of the string to an array. In order to put character to int. we could do s.charAt(i) - 'a' if it was 'a' then it will be zero and 'b' would be 1 etc. very smart.  Look at https://leetcode.com/problems/valid-anagram/solution/ if you are interested for more details